// generated/src/app/api-models/check-print-detail-work.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CheckPrintDetailWork {
  seqCkprtId: number;
  seqApTrans: number;
  seqVendId: number;
  seqVendAddress: number;
  apStatus: string;
  netAmt: number;
  negCheckFlag: string;
  claimStatus: string;
  printFlag: string;
  debitGlNumber1: string;
  fileType: string;
  checkNumber: string;
  bankAccountCode: string;
  interestAmt: number;
  discountAmt: number;
  paidNetAmt: number;
  seqRuleId: number;
  uncleanDays: number;
  penaltyAmt: number;
  intPenDscRsn: string;
  penUncleanDays: number;
  offsetFlag: string;
  checkEftAmount: number;
  seqVendAdvPayAccDtl: number;
}